+++ src/bench/bls.cpp
@@ -5,7 +5,7 @@
 
 #include "bench.h"
 #include "random.h"
-#include "crypto/bls_worker.h"
+#include "bls/bls_worker.h"
 #include "utiltime.h"
 
 #include <iostream>
