+++ src/init.cpp
@@ -17,12 +17,12 @@
 #include "activemasternode.h"
 #include "addrman.h"
 #include "amount.h"
+#include "bls/bls.h"
 #include "budget/budgetdb.h"
 #include "budget/budgetmanager.h"
 #include "checkpoints.h"
 #include "compat/sanity.h"
 #include "consensus/upgrades.h"
-#include "crypto/bls.h"
 #include "evo/evonotificationinterface.h"
 #include "fs.h"
 #include "httpserver.h"
